requirements.txt
annotated-types==0.7.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cors==1.0.1
dnspython==2.8.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
future==1.0.0
gevent==25.9.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
py4j==0.10.9.5
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
pymongo==4.15.2
PySocks==1.7.1
pyspark==3.2.2
python-dotenv==1.1.1
requests==2.32.5
requests-file==2.1.0
tldextract==5.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zope.event==6.0
zope.interface==8.0.1
